+++ src/simulation/elements/PLUT.cpp
@@ -48,7 +48,7 @@ Element_PLUT::Element_PLUT()
 //#TPT-Directive ElementHeader Element_PLUT static int update(UPDATE_FUNC_ARGS)
 int Element_PLUT::update(UPDATE_FUNC_ARGS)
 {
-	if (!(random_gen()%100) && ((int)(5.0f*sim->pv[y/CELL][x/CELL]))>(random_gen()%1000))
+	if (!(random_gen()%100) && (5.0f*sim->pv[y/CELL][x/CELL]) > (random_gen()%1000))
 	{
 		sim->create_part(i, x, y, PT_NEUT);
 	}
